refactor: abstract canister id
