+++ packages/sns/src/governance.canister.ts
@@ -6,21 +6,14 @@ import type {
 } from "../candid/sns_governance";
 import { idlFactory as certifiedIdlFactory } from "../candid/sns_governance.certified.idl";
 import { idlFactory } from "../candid/sns_governance.idl";
+import { Canister } from "./services/canister";
 import { MAX_LIST_NEURONS_RESULTS } from "./constants/governance.constants";
-import type { Canister } from "./types/canister";
 import type { CanisterOptions } from "./types/canister.options";
-import type { CanisterParams } from "./types/canister.params";
 import type { ListNeuronsParams } from "./types/governance.params";
 import { createServices } from "./utils/actor.utils";
 import { toNullable } from "./utils/did.utils";
 
-export class GovernanceCanister implements Canister {
-  private constructor(
-    private readonly governanceCanisterId: Principal,
-    private readonly service: SnsGovernanceCanister,
-    private readonly certifiedService: SnsGovernanceCanister
-  ) {}
-
+export class GovernanceCanister extends Canister<SnsGovernanceCanister> {
   public static create(options: CanisterOptions<SnsGovernanceCanister>) {
     const { service, certifiedService, canisterId } =
       createServices<SnsGovernanceCanister>({
@@ -30,10 +23,6 @@ export class GovernanceCanister implements Canister {
       });
 
     return new GovernanceCanister(canisterId, service, certifiedService);
-  }
-
-  get canisterId(): Principal {
-    return this.governanceCanisterId;
   }
 
   /**
@@ -54,9 +43,4 @@ export class GovernanceCanister implements Canister {
     });
     return neurons;
   };
-
-  private caller = ({
-    certified = true,
-  }: CanisterParams): SnsGovernanceCanister =>
-    certified ? this.certifiedService : this.service;
 }
