--- packages/sns/src/types/canister.ts
@@ -1,5 +1,0 @@
-import type {Principal} from '@dfinity/principal';
-
-export interface Canister {
-    get canisterId(): Principal;
-}
